refactor(blueprint): Add LoadData function
The LoadData function takes "raw" data from a template and loads it similar to LoadConfig.

OCI source info is also introduced as a parameter. This is key, as LoadData and process data can also handle augmenting a blueprint loaded from a source OCI repo, ensuring the OCI source is included on the blueprint and populating its components with proper `source` references.