refactor(template): Make jsonnet template aware of context folder structure
Consolidates the jsonnet templating in to a less abstract component. Now, it has some awareness of the files in the `contexts/_template` structure. This facilitates internal discovery in which we selectively render patch templates that are referenced in the `blueprint.jsonnet`.

It also made sense to move all kustomize template processing in to the `kustomize/` subfolder.