fix(config): Ensure proper config loading order
A default configuration was loaded on top of the local `windsor.yaml` config. This caused incorrect template values on running `windsor init`.

A few areas contributed to this. To remediate, the `--reset` logic that prevented file config loading has been removed as it is no longer the desired functionality. Furthermore, we ensure that we only load a default configuration if an existing context config has not loaded.