feature(cmd): Added windsor build-id command
The `windsor build-id` command is used to manage an incremental build ID for local development purposes. This build ID always increments as it is generated from the unix epoch time stamp. The `windsor build-id --new` flag triggers the creation of a new build ID.

The build ID is stored in a `.windsor/.build-id` file. It is exported as the `BUILD_ID` env var, as well as made available throughout the cluster configuration using the ${BUILD_ID} post build variable.