fix(docker): Workaround secure hardening feature in docker v28
In docker v28, a security hardening feature broke Windsor's ability to route traffic through the VM to the containers. The feature drops packets originating from a non-host. Since we intend for packets to originate from the host interface, we can add this as trusted using new Docker network options in the docker-compose config.

The solution is to add:

```
driver_opts:
  com.docker.network.bridge.trusted_host_interfaces: "col0"
```

When the running docker engine is >= v28 and we're using Colima.

Security hardening outlined in https://www.docker.com/blog/docker-engine-28-hardening-container-networking-by-default/