refactor(config): Begin designing v1alpha2 config schema
The v1alpha2 schema represents the next structure of the `windsor.yaml` file. The major change involves creating a dedicated block for configuring the workstation. Now, workstations could be considered more of a common resource that does not strictly belong to a context. If a local workstation exists, contexts can provision resources on it and leverage its services.

Other changes involve organizing cloud provider blocks, and simplifying Terraform backend config.

Some minor changes like dropping "provider" and "id" are in preparation for a blueprint focused `values.yaml` feature in the future.