refactor(blueprint): Template and apply kustomize values and patches in memory
The templating architecture is shifting. Values and patches placed in `contexts/<context>/kustomize` are now treated as overrides.

As such, `windsor init` will no longer output templated values to these folders, and instead, handle applying them in-memory. The implementation then applies user overrides on top of the base in-memory values before applying to the cluster.