enhancement(blueprint): Hide templated content
If content is templated, it will remain hidden and not generated to `contexts/<context>/...`. This includes both `terraform/` and `patches/`. These are now written to hidden folders or handled in-memory only. This change makes it so that the user only needs to concern themselves with overrides, and the tool is less focused on "templating" as its means of operation.

Signed-off-by: Ryan VanGundy <85766511+rmvangun@users.noreply.github.com>